+++ MatrixSDK/Crypto/Algorithms/Megolm/MXMegolmEncryption.m
@@ -193,7 +193,7 @@
     // with them, which means that they will have announced any new devices via
     // an m.new_device.
     MXWeakify(self);
-    return [crypto.deviceList downloadKeys:users forceDownload:NO success:^(MXUsersDevicesMap<MXDeviceInfo *> *devices) {
+    return [crypto.deviceList downloadKeys:users forceDownload:NO success:^(MXUsersDevicesMap<MXDeviceInfo *> *devices, NSDictionary<NSString *,MXCrossSigningInfo *> *crossSigningKeysMap) {
         MXStrongifyAndReturnIfNil(self);
 
         BOOL encryptToVerifiedDevicesOnly = self->crypto.globalBlacklistUnverifiedDevices